# Add string-based CallSite variant to wrapper enums